refactor: move most of the AST elements to their own file (#543)
And a tiny test optimization
